feat: [UF-189] 헤더, 필터, 목록, 액션 핸들러로 책임 분리 #165
